requirements.txt
asgiref==3.5.2
Django==3.1.6
djangorestframework==3.12.4
python-environ==0.4.54
pytz==2022.2.1
sqlparse==0.4.2
